Struct trillium_channels::ChannelConn

pub struct ChannelConn<'a> { /* private fields */ }

A ChannelConn is a wrapper around a WebSocketConn that also contains a ChannelClient

It that provides convenient access to functions from the ChannelClient held in the WebSocketConn’s StateSet, and dereferences to the WebSocketConn for other functionality.

Implementations

impl ChannelConn<'_>

pub fn client(&self) -> Option<&ChannelClient>

Borrow the channel client

pub fn conn(&self) -> &WebSocketConn

Borrow the websocket conn

pub fn broadcast(&self, event: impl Into<ChannelEvent>)

Send a ChannelEvent to all connected clients. Note that these messages will only reach clients that subscribe to the event’s topic.

pub async fn send_event(&self, event: impl Into<ChannelEvent>)

Send a ChannelEvent to this specific client. Note that this message will only be received if the client subscribes to the event’s topic.

pub async fn reply_ok(&self, event: &ChannelEvent, response: &impl Serialize)

Send an ok reply in reference to the provided ChannelEvent with the provided response payload.

Note that this sets the event as "phx_reply" and the payload as {"status": "ok", "response": response }, as well as setting the reference field.

pub async fn reply_error(&self, event: &ChannelEvent, error: &impl Serialize)

Send an error reply in reference to the provided ChannelEvent with the provided response payload.

Note that this sets the event as "phx_error" as well as setting the reference field.

pub async fn allow_join(&self, event: &ChannelEvent, value: &impl Serialize)

Join a topic, sending an ok reply with the provided optional value. This sends an ok reply to the client as well as adding the topic to the client’s subscriptions.

pub async fn allow_leave(&self, event: &ChannelEvent, payload: &impl Serialize)

Leave a topic as requested by the provided channel event, including the optional payload. This sends an ok reply to the client as well as removing the channel from the client’s subscriptions.